[back-end] move source into src/, API spec into api/
the fastify-openapi-glue generated openApi.json file can be removed in
favour of the openapi generated
SWAGGERHUBWULEE-mykomap-0.1.0-resolved.json, with a shorter name.
